+++ NEWS.md
@@ -4,6 +4,7 @@ Red Panda C++ Version 2.11
   - enhancement: Auto suggest keyword "operator" when define functions.
   - enhancement: Differentiate class and constructors in syntax color and jupming to declarations.
   - enhancement: Improve parsing for operator overloading.
+  - fix: Parser can't correctly differentiate function and var initialization.
 
 Red Panda C++ Version 2.10